refactor: update folder structure
Extract each type and enum into its own file to improve maintainability.
