+++ src/lib/enums/fileMode.ts
@@ -1,0 +1,4 @@
+export enum FileMode {
+  Manual = 'manual',
+  Automatic = 'automatic',
+}
