+++ src/lib/enums/fileType.ts
@@ -1,0 +1,5 @@
+export enum FileType {
+  Json = 'json',
+  Po = 'po',
+  Code = 'code',
+}
