+++ src/lib/enums/logVerbosity.ts
@@ -1,0 +1,4 @@
+export enum LogVerbosity {
+  Verbose,
+  Important,
+}
