+++ src/lib/enums/statusBarItemType.ts
@@ -1,0 +1,5 @@
+export enum StatusBarItemType {
+  PO,
+  JSON,
+  CODE,
+}
