+++ src/lib/statusBarManager.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { StatusBarItemType } from './Enums';
+import { StatusBarItemType } from './enums/statusBarItemType';
 
 class StatusBarManager {
   private static _instance: StatusBarManager;
