+++ src/lib/types/callbackOnMatch.ts
@@ -1,0 +1,1 @@
+export type CallbackOnMatch = (output: string) => void;
